fix: 스크랩 기능 재구현

+++ src/main/java/hanium/dtc/community/service/CommentService.java
@@ -2,6 +2,7 @@ package hanium.dtc.community.service;
 
 import hanium.dtc.community.domain.Comment;
 import hanium.dtc.community.domain.Post;
+import hanium.dtc.user.domain.User;
 import hanium.dtc.community.dto.request.CommentRequest;
 import hanium.dtc.community.dto.response.CommentResponse;
 import hanium.dtc.community.repository.CommentRepository;
@@ -28,10 +29,14 @@ public class CommentService {
     private final UserRepository userRepository;
 
     @Transactional
-    public boolean createComment(Long postId, CommentRequest commentRequest) {
+    public boolean createComment(Long postId, Long userId, CommentRequest commentRequest) {
+
 
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new CommonException(ErrorCode.NOT_FOUND_POST));
+
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new CommonException(ErrorCode.NOT_FOUND_USER));
 
         Comment comment = new Comment(
                 commentRequest.content(),
@@ -39,14 +44,23 @@ public class CommentService {
                 commentRequest.commentId(),
                 post
         );
+        comment.setUser(user);
         commentRepository.save(comment);
         return true;
     }
 
     @Transactional
-    public boolean updateComment(Long commentId, CommentRequest commentRequest) {
+    public boolean updateComment(Long postId, Long commentId, Long userId, CommentRequest commentRequest) {
+
+        Post post = postRepository.findById(postId)
+                .orElseThrow(() -> new CommonException(ErrorCode.NOT_FOUND_POST));
+
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new CommonException(ErrorCode.NOT_FOUND_COMMENT));
+
+        if (!comment.getUser().getId().equals(userId)) {
+            throw new CommonException(ErrorCode.FORBIDDEN_ROLE);
+        }
 
         comment.updateContent(commentRequest.content());
         commentRepository.save(comment);
@@ -54,9 +68,17 @@ public class CommentService {
     }
 
     @Transactional
-    public boolean deleteComment(Long commentId) {
+    public boolean deleteComment(Long postId, Long commentId, Long userId) {
+
+        Post post = postRepository.findById(postId)
+                .orElseThrow(() -> new CommonException(ErrorCode.NOT_FOUND_POST));
+
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new CommonException(ErrorCode.NOT_FOUND_COMMENT));
+
+        if (!comment.getUser().getId().equals(userId)) {
+            throw new CommonException(ErrorCode.FORBIDDEN_ROLE);
+        }
 
         commentRepository.delete(comment);
         return true;
@@ -79,10 +101,7 @@ public class CommentService {
 
             CommentResponse response = CommentResponse.builder()
                     .content(comment.getContent())
-                    .userCommentResponse(UserCommentResponse.builder()
-                            .nickname(comment.getUser().getNickname())
-                            .image(comment.getUser().getImage())
-                            .build())
+                    .user(userCommentResponse)
                     .like(comment.getCommentLike())
                     .commentTime(comment.getCommentTime())
                     .isReply(comment.getIsReply())
